API Separate and rename NewsHolderController
